Reserves a new parent-controlled fuse that allows owners of subnames to extend their own expiry.
This gives a first-class easy way to allow perpetual "unruggable" renewals for subname owners, without needing to use some other contract to indirectly call setChildFuses.

This will not be burned on .eth 2LDs, and it doesn't matter there anyway since the expiry will always inherit from the registrar in that case.